+++ src/main/java/com/example/restaurantreservationaa/domain/dto/menuitem/MenuItemRegistrationDto.java
@@ -11,7 +11,7 @@ import lombok.NoArgsConstructor;
 public class MenuItemRegistrationDto {
     private String name;
     private String description;
-    private double price;
+    private Float price;
     private String category;
     private Boolean isVegetarian;
 }
