+++ src/main/java/com/example/restaurantreservationaa/domain/dto/order/OrderOutDto.java
@@ -18,5 +18,5 @@ public class OrderOutDto {
     private List<Beverage> beverages;
     private String status;
     private Date orderDate;
-    private double totalPrice;
+    private Float totalPrice;
 }
