+++ src/main/java/com/example/restaurantreservationaa/domain/dto/order/OrderRegistrationDto.java
@@ -17,5 +17,5 @@ public class OrderRegistrationDto {
     private List<Beverage> beverages;
     private String status;
     private Date orderDate;
-    private double totalPrice;
+    private Float totalPrice;
 }
